Feat: Full Text Search by Title
